permutations use u32 not usize

+++ src/protocol/reveal.rs
@@ -77,7 +77,7 @@ pub async fn reveal_malicious<F: Field, G: Field>(
 pub async fn reveal_permutation<F: Field>(
     ctx: ProtocolContext<'_, Replicated<F>, F>,
     permutation: &[Replicated<F>],
-) -> Result<Vec<usize>, BoxError> {
+) -> Result<Vec<u32>, BoxError> {
     let revealed_permutation = try_join_all(zip(repeat(ctx), permutation).enumerate().map(
         |(index, (ctx, input))| async move {
             let reveal_value = reveal(ctx, RecordId::from(index), *input).await;
